+++ tests/Definition/Plugin/EndpointsDefinitionPluginTest.php
@@ -21,7 +21,7 @@ use Ynlo\GraphQLBundle\Model\DeleteNodePayload;
 use Ynlo\GraphQLBundle\Model\OrderBy;
 use Ynlo\GraphQLBundle\Model\PageInfo;
 use Ynlo\GraphQLBundle\Model\UpdateNodePayload;
-use Ynlo\GraphQLBundle\Tests\Fixtures\AppBundle\Entity\Comment;
+use Ynlo\GraphQLBundle\Tests\Fixtures\AppBundle\Entity\PostComment;
 use Ynlo\GraphQLBundle\Tests\Fixtures\AppBundle\Entity\Post;
 use Ynlo\GraphQLBundle\Tests\Fixtures\AppBundle\Entity\User;
 use Ynlo\GraphQLBundle\Tests\Fixtures\AppBundle\Model\HasAuthorInterface;
@@ -75,7 +75,7 @@ class EndpointsDefinitionPluginTest extends TestCase
         TestDefinitionHelper::loadAnnotationDefinitions(UpdateNodePayload::class, $endpoint);
         TestDefinitionHelper::loadAnnotationDefinitions(DeleteNodePayload::class, $endpoint);
         TestDefinitionHelper::loadAnnotationDefinitions(Post::class, $endpoint);
-        TestDefinitionHelper::loadAnnotationDefinitions(Comment::class, $endpoint);
+        TestDefinitionHelper::loadAnnotationDefinitions(PostComment::class, $endpoint);
         TestDefinitionHelper::loadAnnotationDefinitions(User::class, $endpoint);
         TestDefinitionHelper::loadAnnotationDefinitions(OrderBy::class, $endpoint);
         TestDefinitionHelper::loadAnnotationDefinitions(PageInfo::class, $endpoint);